+++ test/unit/test_deployments.py
@@ -24,11 +24,14 @@ def mock_config(deployment_type='deploy', prod_compose=False):
     return mock_config
 
 
+@patch('harvey.locks.Lock.lookup_project_lock', return_value=False)
 @patch('harvey.config.Config.use_slack', True)
 @patch('harvey.git.Git.update_git_repo')
 @patch('harvey.deployments.Deployment.open_project_config', return_value=mock_config())
 @patch('harvey.messages.Message.send_slack_message')
-def test_initialize_deployment_slack(mock_slack_message, mock_open_project_config, mock_update_git_repo, mock_webhook):
+def test_initialize_deployment_slack(
+    mock_slack_message, mock_open_project_config, mock_update_git_repo, mock_project_lock, mock_webhook
+):
     _, _, _ = Deployment.initialize_deployment(mock_webhook)
 
     mock_slack_message.assert_called_once()