+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - API has more unified error handling for 5xx error types
 - Introduced new HarveyError and cleaned up various error handling throughout the app. All messages should now properly get logged when a deployment gets killed, errors raised when Sentries need to be triggered, and various log levels were corrected
+- Healthchecks of containers now check when the container was started to ensure that not only are they running, but that they restarted within the last 60 seconds as a part of the deploy
 - If a project's lock status cannot be determined, we now kill the process instead of continuing and logging only
 - Wraps `store_deployment_details` in a try/except to log out errors with saving details to the DB